Add support to include Gazebo in  idjl-software-dependencies-vcpkg
On the top of the usual `vcpkg-idjl.zip` archive, this commit also generates the archive `vcpkg-idjl-with-gazebo.zip`,
that is like the usual `vcpkg-idjl.zip` archive, but that also contains a `C:\idjl\gazebo\install` directory in which gazebo 11
and its dependencies are installed.

As Gazebo in this case it is  not installed via vcpkg, it is not possible to find it via the official vcpkg's integration via the
`CMAKE_TOOLCHAIN_FILE`, so also a set of scripts is provided to set for a given terminal (`setup-deps.bat` and `setup-deps.sh`) or
for the system (`addPathsToUserEnvVariables-deps.ps1` and `removePathsFromUserEnvVariables-deps.ps1`) the necessary environment
variables to find the dependencies and run Gazebo.

The specific version of Gazebo installed by the scripts are contained in the gazebo-repos.yaml file. They do refer to specific
commit due to the fact that the Gazebo windows support requires some of the latest fix (some still pending, see https://github.com/osrf/gazebo/pull/2719),
but they will be updated to tags once a release will be made including the required tags.